Move util files from ether-email-auth to utils repo
