+++ packages/relayer/Cargo.toml
@@ -24,7 +24,8 @@ serde_json = "1.0.68"
 tiny_http = "0.12.0"
 lettre = { version = "0.10.4", features = ["tokio1", "tokio1-native-tls"] }
 ethers = { version = "2.0.10", features = ["abigen"] }
-relayer-utils = { version = "0.3.7", git = "https://github.com/zkemail/relayer-utils.git" }
+# relayer-utils = { version = "0.3.7", git = "https://github.com/zkemail/relayer-utils.git" }
+relayer-utils = { path = "../../../relayer-utils" }
 futures = "0.3.28"
 sqlx = { version = "=0.7.3", features = ["postgres", "runtime-tokio"] }
 regex = "1.10.2"
