+++ packages/relayer/src/chain.rs
@@ -4,6 +4,7 @@ use ethers::middleware::Middleware;
 use ethers::prelude::*;
 use ethers::signers::Signer;
 use relayer_utils::converters::u64_to_u8_array_32;
+use relayer_utils::TemplateValue;
 
 const CONFIRMATIONS: usize = 1;
