--- packages/relayer/src/utils/mod.rs
@@ -1,10 +1,0 @@
-// TODO: Remove this once we remove the utils
-#![allow(clippy::module_inception)]
-
-pub mod strings;
-pub mod subject_templates;
-pub mod utils;
-
-pub use strings::*;
-pub use subject_templates::*;
-pub use utils::*;
