--- packages/relayer/src/utils/utils.rs
@@ -1,94 +1,0 @@
-#![allow(clippy::upper_case_acronyms)]
-#![allow(clippy::identity_op)]
-
-use crate::*;
-use ethers::abi::Token;
-use ethers::types::{Bytes, U256};
-
-use ::serde::Deserialize;
-
-use relayer_utils::*;
-use std::collections::hash_map::DefaultHasher;
-use std::hash::{Hash, Hasher};
-
-#[derive(Debug, Clone, Deserialize)]
-pub struct ProverRes {
-    proof: ProofJson,
-    pub_signals: Vec<String>,
-}
-
-#[derive(Debug, Clone, Deserialize)]
-pub struct ProofJson {
-    pi_a: Vec<String>,
-    pi_b: Vec<Vec<String>>,
-    pi_c: Vec<String>,
-}
-
-impl ProofJson {
-    pub fn to_eth_bytes(&self) -> Result<Bytes> {
-        let pi_a = Token::FixedArray(vec![
-            Token::Uint(U256::from_dec_str(self.pi_a[0].as_str())?),
-            Token::Uint(U256::from_dec_str(self.pi_a[1].as_str())?),
-        ]);
-        let pi_b = Token::FixedArray(vec![
-            Token::FixedArray(vec![
-                Token::Uint(U256::from_dec_str(self.pi_b[0][1].as_str())?),
-                Token::Uint(U256::from_dec_str(self.pi_b[0][0].as_str())?),
-            ]),
-            Token::FixedArray(vec![
-                Token::Uint(U256::from_dec_str(self.pi_b[1][1].as_str())?),
-                Token::Uint(U256::from_dec_str(self.pi_b[1][0].as_str())?),
-            ]),
-        ]);
-        let pi_c = Token::FixedArray(vec![
-            Token::Uint(U256::from_dec_str(self.pi_c[0].as_str())?),
-            Token::Uint(U256::from_dec_str(self.pi_c[1].as_str())?),
-        ]);
-        Ok(Bytes::from(abi::encode(&[pi_a, pi_b, pi_c])))
-    }
-}
-
-pub async fn generate_proof(
-    input: &str,
-    request: &str,
-    address: &str,
-) -> Result<(Bytes, Vec<U256>)> {
-    let client = reqwest::Client::new();
-    info!(LOG, "prover input {}", input);
-    let res = client
-        .post(format!("{}/prove/{}", address, request))
-        .json(&serde_json::json!({ "input": input }))
-        .send()
-        .await?
-        .error_for_status()?;
-    let res_json = res.json::<ProverRes>().await?;
-    info!(LOG, "prover response {:?}", res_json);
-    let proof = res_json.proof.to_eth_bytes()?;
-    let pub_signals = res_json
-        .pub_signals
-        .into_iter()
-        .map(|str| U256::from_dec_str(&str).expect("pub signal should be u256"))
-        .collect();
-    Ok((proof, pub_signals))
-}
-
-pub fn calculate_default_hash(input: &str) -> String {
-    let mut hasher = DefaultHasher::new();
-    input.hash(&mut hasher);
-    let hash_code = hasher.finish();
-
-    hash_code.to_string()
-}
-
-pub fn calculate_account_salt(email_addr: &str, account_code: &str) -> String {
-    let padded_email_addr = PaddedEmailAddr::from_email_addr(email_addr);
-    let account_code = if account_code.starts_with("0x") {
-        hex_to_field(account_code).unwrap()
-    } else {
-        hex_to_field(&format!("0x{}", account_code)).unwrap()
-    };
-    let account_code = AccountCode::from(account_code);
-    let account_salt = AccountSalt::new(&padded_email_addr, account_code).unwrap();
-
-    field_to_hex(&account_salt.0)
-}
